requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.8.0
celery==5.6.2
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
diff-match-patch==20241021
distro==1.9.0
dj-database-url==3.0.1
Django==6.0
django-axes==8.1.0
django-cors-headers==4.9.0
django-csp==4.0
django-environ==0.12.0
django-import-export==4.4.0
django-ninja==1.5.3
django-redis==6.0.0
django-unfold==0.75.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jiter==0.12.0
kombu==5.6.2
msgpack==1.1.2
nh3==0.3.2
openai==2.14.0
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
psycopg==3.3.2
psycopg-binary==3.3.2
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
tablib==3.9.0
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
whitenoise==6.11.0
zope.interface==8.1.1
